requirements.txt
Flask==3.0.0
Flask-CORS==4.0.0
Flask-Login==0.6.3
mysql-connector-python==8.2.0
bcrypt==4.1.2
python-dotenv==1.0.0
pyinstaller==6.3.0
